+++ R/simOUJ.R
@@ -1,6 +1,6 @@
 #' OUJ process simulation
 #' @description Simulates a Ornstein–Uhlenbeck process with Jumps
-#' @param nsims numble of simulations. Defaults to 2
+#' @param nsims number of simulations. Defaults to 2
 #' @param S0 S at t=0
 #' @param mu Mean reversion level
 #' @param theta Mean reversion speed